Moved logic from component to service
